A bit of refactoring to move the useEffect up to be able to pass to both booking and user profile components.
